+++ tunnels/EncryptionClient/include/EncryptionClient/structure.h
@@ -43,7 +43,7 @@ WW_EXPORT void         encryptionclientTunnelDestroy(tunnel_t *t);
 WW_EXPORT tunnel_t    *encryptionclientTunnelCreate(node_t *node);
 WW_EXPORT api_result_t encryptionclientTunnelApi(tunnel_t *instance, sbuf_t *message);
 
-void encryptionclientTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset);
+void encryptionclientTunnelOnIndex(tunnel_t *t, uint16_t index, uint32_t *mem_offset);
 void encryptionclientTunnelOnChain(tunnel_t *t, tunnel_chain_t *chain);
 void encryptionclientTunnelOnPrepair(tunnel_t *t);
 void encryptionclientTunnelOnStart(tunnel_t *t);
