+++ tunnels/Socks5Server/include/Socks5Server/structure.h
@@ -119,7 +119,7 @@ WW_EXPORT void         socks5serverTunnelDestroy(tunnel_t *t);
 WW_EXPORT tunnel_t    *socks5serverTunnelCreate(node_t *node);
 WW_EXPORT api_result_t socks5serverTunnelApi(tunnel_t *instance, sbuf_t *message);
 
-void socks5serverTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset);
+void socks5serverTunnelOnIndex(tunnel_t *t, uint16_t index, uint32_t *mem_offset);
 void socks5serverTunnelOnChain(tunnel_t *t, tunnel_chain_t *chain);
 void socks5serverTunnelOnPrepair(tunnel_t *t);
 void socks5serverTunnelOnStart(tunnel_t *t);
