+++ tunnels/TcpOverUdpClient/include/TcpOverUdpClient/structure.h
@@ -101,7 +101,7 @@ WW_EXPORT void         tcpoverudpclientTunnelDestroy(tunnel_t *t);
 WW_EXPORT tunnel_t    *tcpoverudpclientTunnelCreate(node_t *node);
 WW_EXPORT api_result_t tcpoverudpclientTunnelApi(tunnel_t *instance, sbuf_t *message);
 
-void tcpoverudpclientTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset);
+void tcpoverudpclientTunnelOnIndex(tunnel_t *t, uint16_t index, uint32_t *mem_offset);
 void tcpoverudpclientTunnelOnChain(tunnel_t *t, tunnel_chain_t *chain);
 void tcpoverudpclientTunnelOnPrepair(tunnel_t *t);
 void tcpoverudpclientTunnelOnStart(tunnel_t *t);
