+++ tunnels/TcpUdpConnector/instance/index.c
@@ -1,6 +1,6 @@
 #include "structure.h"
 
-void tcpudpconnectorTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset)
+void tcpudpconnectorTunnelOnIndex(tunnel_t *t, uint16_t index, uint32_t *mem_offset)
 {
     tunnelDefaultOnIndex(t, index, mem_offset);
 }
